+++ PhotoManager/src/main/java/picto/com/photomanager/global/postDomain/entity/PhotoRecordId.java
@@ -17,7 +17,7 @@ import java.util.Objects;
 @EqualsAndHashCode
 public class PhotoRecordId implements java.io.Serializable {
     private static final long serialVersionUID = 4786054818538112941L;
-    private Integer agentId;
+    private Long agentId;
     private PhotoId photoId;
 
     //@Column(name = "photo_id", nullable = false)
