+++ genenrator/src/main/java/picto/com/generator/global/entity/PhotoRecordId.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @EqualsAndHashCode
 public class PhotoRecordId implements java.io.Serializable {
     private static final long serialVersionUID = 4786054818538112941L;
-    private Integer agentId;
+    private Long agentId;
     private PhotoId photoId;
 }
