+++ genenrator/src/main/java/picto/com/generator/global/repositories/FilterRepository.java
@@ -3,5 +3,5 @@ package picto.com.generator.global.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import picto.com.generator.global.entity.Filter;
 
-public interface FilterRepository extends JpaRepository<Filter, Integer> {
+public interface FilterRepository extends JpaRepository<Filter, Long> {
 }
