+++ genenrator/src/main/java/picto/com/generator/global/repositories/SessionRepository.java
@@ -3,5 +3,5 @@ package picto.com.generator.global.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import picto.com.generator.global.entity.Session;
 
-public interface SessionRepository extends JpaRepository<Session, Integer> {
+public interface SessionRepository extends JpaRepository<Session, Long> {
 }
