+++ genenrator/src/main/java/picto/com/generator/global/repositories/TitleListRepository.java
@@ -3,5 +3,5 @@ package picto.com.generator.global.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import picto.com.generator.global.entity.TitleList;
 
-public interface TitleListRepository extends JpaRepository<TitleList, Integer> {
+public interface TitleListRepository extends JpaRepository<TitleList, Long> {
 }
