+++ genenrator/src/main/java/picto/com/generator/global/repositories/UserSettingRepositroy.java
@@ -3,5 +3,5 @@ package picto.com.generator.global.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import picto.com.generator.global.entity.UserSetting;
 
-public interface UserSettingRepositroy extends JpaRepository<UserSetting, Integer> {
+public interface UserSettingRepositroy extends JpaRepository<UserSetting, Long> {
 }
